chore(batcher): reorder block builder file
