+++ crates/batcher/src/lib.rs
@@ -13,3 +13,4 @@ mod proposal_manager;
 mod proposal_manager_test;
 #[cfg(test)]
 mod test_utils;
+mod transaction_executor;
