+++ crates/batcher/src/transaction_executor.rs
@@ -1,0 +1,41 @@
+use blockifier::blockifier::transaction_executor::{
+    TransactionExecutor,
+    TransactionExecutorResult,
+    VisitedSegmentsMapping,
+};
+use blockifier::bouncer::BouncerWeights;
+use blockifier::state::cached_state::CommitmentStateDiff;
+use blockifier::state::state_api::StateReader;
+use blockifier::transaction::objects::TransactionExecutionInfo;
+use blockifier::transaction::transaction_execution::Transaction as BlockifierTransaction;
+#[cfg(test)]
+use mockall::automock;
+
+#[cfg_attr(test, automock)]
+pub trait TransactionExecutorTrait: Send {
+    fn add_txs_to_block(
+        &mut self,
+        txs: &[BlockifierTransaction],
+    ) -> Vec<TransactionExecutorResult<TransactionExecutionInfo>>;
+    fn close_block(
+        &mut self,
+    ) -> TransactionExecutorResult<(CommitmentStateDiff, VisitedSegmentsMapping, BouncerWeights)>;
+}
+
+impl<S: StateReader + Send + Sync> TransactionExecutorTrait for TransactionExecutor<S> {
+    /// Adds the transactions to the generated block and returns the execution results.
+    fn add_txs_to_block(
+        &mut self,
+        txs: &[BlockifierTransaction],
+    ) -> Vec<TransactionExecutorResult<TransactionExecutionInfo>> {
+        self.execute_txs(txs)
+    }
+    /// Finalizes the block creation and returns the commitment state diff, visited
+    /// segments mapping and bouncer.
+    fn close_block(
+        &mut self,
+    ) -> TransactionExecutorResult<(CommitmentStateDiff, VisitedSegmentsMapping, BouncerWeights)>
+    {
+        self.finalize()
+    }
+}
